fix: workaround for using the app in Blue Cat's Patchwork
For some reason Blue Cat's Patchwork does not except multiple messages in the midi buffer. It will only process some of them with hanging and missing notes as a result. By adding only one message to the buffer this problem is avoided. However, it introduces latency. Therefore the workaround is only usable if the sample buffer is small enough.